requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
pycaret==3.3.2
langfuse>=3.0.0
openai>=1.35.0